Backend/src/requirements.txt
absl-py==1.4.0
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.3.0
astunparse==1.6.3
async-lru==2.0.4
asyncpg==0.30.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
category_encoders==2.8.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cloudpickle==3.1.1
cobra==0.29.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.0
coverage==7.3.1
cvxopt==1.3.2
cycler==0.11.0
daqp==0.5.1
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
depinfo==2.2.0
diskcache==5.6.3
Django==4.2.5
dotenv==0.9.9
ecos==2.0.12
executing==1.2.0
fastapi==0.115.13
fastjsonschema==2.18.0
feature-engine==1.8.3
filelock==3.18.0
findspark==2.0.1
Flask==2.3.2
Flask-API==3.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
fsspec==2025.5.1
funcy==2.0
future==1.0.0
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphql-core==3.2.6
greenlet==3.2.3
grpcio==1.58.0
h11==0.16.0
h5py==3.9.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.4
image==1.5.33
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.15.0
ipython-genutils==0.2.0
ipythonblocks==1.9.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.5.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.44.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
metakernel==0.30.0
mistune==3.0.1
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numba==0.61.2
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
optlang==1.8.3
osqp==0.6.3
overrides==7.4.0
packaging==23.1
pandas==2.3.0
pandocfilters==1.5.0
parso==0.8.3
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
populate==0.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.22
pydantic_core==2.33.2
Pygments==2.16.1
PyMySQL==1.1.1
pyparsing==3.0.9
pyreadline3==3.5.4
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
python-libsbml==5.20.5
python-multipart==0.0.20
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qdldl==0.1.7.post0
qpsolvers==4.0.0
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.10.3
rsa==4.9
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.11.2
scs==3.2.3
Send2Trash==1.8.2
shap==0.48.0
six==1.16.0
slicer==0.0.8
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spylon==0.3.0
spylon-kernel==0.4.1
SQLAlchemy==2.0.41
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.46.2
statsmodels==0.14.4
strawberry-graphql==0.275.2
swiglpk==5.0.12
sympy==1.14.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.6.0
tinycss2==1.2.1
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.9.0
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.34.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.6
widgetsnbextension==4.0.9
wrapt==1.15.0
